Move the create_xform_version function into logger_tools.py
- Handle the IntegrityError raised when creating a duplicate
XFormVersion
- Move functionality responsible for the creation of the XFormVersion
into the publish functions
